ExtractEmails/requirements.txt
# Python Email Extraction Tool
# Requires Python 3.9 or higher

requests>=2.31.0
beautifulsoup4>=4.12.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
openai>=1.0.0